src/AdvancedOperations.cs

using System;
using System.Collections.Generic;

public class AdvancedOperations {
    public AdvancedOperations() {
        // Constructor code if needed
    }

    // Method to find the maximum value in an array
    public int FindMax(int[] arr) {
        // Method implementation removed
    }

    // Method to reverse a string
    public string ReverseString(string str) {
        // Method implementation removed
    }

    // Method to format a date as 'YYYY-MM-DD'
    public string FormatDate(DateTime date) {
        // Method implementation removed
    }

    // Method to calculate the factorial of a number
    public int Factorial(int n) {
        // Method implementation removed
    }

    // Method to find the nth Fibonacci number
    public int Fibonacci(int n) {
        // Method implementation removed
    }

    // Method to check if a string is a palindrome
    public bool IsPalindrome(string str) {
        // Method implementation removed
    }

    // Method to merge and sort two arrays
    public int[] MergeAndSort(int[] arr1, int[] arr2) {
        // Method implementation removed
    }

    /**
     * Calculates statistics for a set of numbers.
     * @param nums - The set of numbers.
     * @return The statistics including sum, average, min, and max.
     */
    public Dictionary<string, object> CalculateStatistics(int[] nums) {
        var stats = new Dictionary<string, object>();
        if (nums.Length == 0) {
            stats["sum"] = 0;
            stats["average"] = double.NaN;
            stats["min"] = null;
            stats["max"] = null;
            return stats;
        }

        int sum = 0;
        int min = nums[0];
        int max = nums[0];
        foreach (int num in nums) {
            sum += num;
            if (num < min) min = num;
            if (num > max) max = num;
        }
        double average = (double) sum / nums.Length;

        stats["sum"] = sum;
        stats["average"] = average;
        stats["min"] = min;
        stats["max"] = max;
        return stats;
    }
}